no_defausse

+++ projet/src/loveletters/controler/ControlerPartie.php
@@ -200,7 +200,6 @@ class ControlerPartie {
       // Defausse des 3 premières cartes
       $defausse = new Defausse;
       $defausse->idManche=$manche->idManche;
-      $defausse->idJoueur=null;
       $defausse->save();
       for($i = 0; $i<3; $i++){
         $randomId=$cartes[array_rand($cartes, 1)];
@@ -241,11 +240,13 @@ class ControlerPartie {
       $possede = Possede::where('idJoueur','=',$_SESSION['idJoueur'])
                        ->where('idCarte','=',$idCarte)
                        ->first();
-      if($possede->nbCartes>1){
-        $possede->nbCartes--;
-        $possede->save();
-      }else{
-        $possede->forceDelete();
+      if($possede!=null){
+        if($possede->nbCartes>1){
+          $possede->nbCartes--;
+          $possede->save();
+        }else{
+          $possede->forceDelete();
+        }
       }
       // Pose la carte sur le terrain (du coté du joueur)
       $posee = new Posee;
@@ -255,10 +256,6 @@ class ControlerPartie {
       // Applique l'effet de la carte
       switch($idCarte){
         case 4: $this->protect();
-                break;
-        case 5:
-                break;
-        case 6:
                 break;
       }
       $joueur->etat_tour='fin';
@@ -299,10 +296,10 @@ class ControlerPartie {
 
   public function effetBaron($idJoueur){
     DBConnection::getInstance();
-    $posees= Posee::where('idJoueur','=',$idJoueur)->get();
+    $possedes= Possede::where('idJoueur','=',$idJoueur)->get();
     $cartes=array();
-    foreach($posees as $posee){
-      $cartes[]= Carte::where('idCarte','=',$posee->idCarte)->first();
+    foreach($possedes as $possede){
+      $cartes[]= Carte::where('idCarte','=',$possede->idCarte)->first();
     }
     $vueCarte = new VueCarte();
     echo $vueCarte->cardsToHtml($cartes);
@@ -327,33 +324,11 @@ class ControlerPartie {
     if(!$joueur->protect){
       $possede = Possede::where('idJoueur','=',$idJoueur)->first();
       $idManche = Participe::where('idJoueur','=',$_SESSION['idJoueur'])->max('idManche');
-      $defausse = Defausse::where('idJoueur','=',$_SESSION['idJoueur'])
-                          ->where('idManche','=',$idManche)
-                          ->first();
-      if($defausse==null){
-        $defausse = new Defausse;
-        $defausse->idJoueur=$_SESSION['idJoueur'];
-        $defausse->idManche=$idManche;
-        $defausse->save();
-      }
-      if($possede!=null){
-        $estPlacee = EstPlacee::where('idCarte','=',$possede->idCarte)
-                              ->where('idDefausse','=',$defausse->idDefausse)
-                              ->first();
-        if($estPlacee==null){
-          $estPlacee = new EstPlacee;
-          $estPlacee->idCarte=$possede->idCarte;
-          $estPlacee->idDefausse=$defausse->idDefausse;
-          $estPlacee->nbCartes=1;
-        }else{
-          $estPlacee->nbCartes=$estPlacee->nbCartes+1;
-        }
-        $estPlacee->save();
-        $possede->forceDelete();
-        if($estPlacee->idCarte == 8){
-          $this->eliminer($idJoueur);
-        }
-      }
+      $posee = new Posee;
+      $posee->idJoueur=$idJoueur;
+      $posee->idCarte=$possede->idCarte;
+      $posee->save();
+      $possede->forceDelete();
       $this->pioche($idJoueur,true);
     }
   }
@@ -437,33 +412,21 @@ class ControlerPartie {
       $possedes= Possede::where('idJoueur','=',$participe->idJoueur)->get();
       $joueurs[$participe->idJoueur]=0;
       foreach($possedes as $possede){
-        $joueurs[$participe->idJoueur]++;
+        for($i = 0;$i<$possede->nbCartes;$i++){
+          $joueurs[$participe->idJoueur]++;
+        }
       }
     }
     echo json_encode($joueurs);
   }
 
-  public function affichageDefausse($idJoueur=null){
+  public function affichageDefausse(){
     DBConnection::getInstance();
     if(!isset($_SESSION)){
       session_start();
     }
     $idManche=Manche::where('idPartie','=',$_SESSION['idPartie'])->max('idManche');
-    if($idJoueur!=null){
-      $defausse = Defausse::where('idManche','=',$idManche)
-                          ->where('idJoueur','=',$idJoueur)
-                          ->first();
-    }else{
-      $defausse = Defausse::where('idManche','=',$idManche)
-                          ->where('idJoueur','=',null)
-                          ->first();
-    }
-    if($defausse==null){
-      $defausse = new Defausse;
-      $defausse->idJoueur=$idJoueur;
-      $defausse->idManche=$idManche;
-      $defausse->save();
-    }
+    $defausse = Defausse::where('idManche','=',$idManche)->first();
     $estPlacees = EstPlacee::where('idDefausse','=',$defausse->idDefausse)->get();
     $cartes= array();
     foreach($estPlacees as $estPlacee){
@@ -477,31 +440,25 @@ class ControlerPartie {
     echo $res;
   }
 
-  public function defausse($idCarte, $idJoueur=null, $selecteur=null){
+  public function defausse($idCarte, $idJoueur=null){
     DBConnection::getInstance();
     if(!isset($_SESSION)){
       session_start();
     }
     $idManche=Manche::where('idPartie','=',$_SESSION['idPartie'])->max('idManche');
     if($idJoueur!=null){
-      $defausse = Defausse::where('idManche','=',$idManche)
-                          ->where('idJoueur','=',$idJoueur)
-                          ->first();
-      if($selecteur=='main'){
-        $possede = Possede::where('idCarte','=',$idCarte)
-                          ->where('idJoueur','=',$idJoueur)
-                          ->first();
-        if($possede->nbCartes==1){
-          $possede->forceDelete();
-        }else{
-          $possede->nbCartes=$possede->nbCartes-1;
-          $possede->save();
-        }
+      $posee = new Posee;
+      $posee->idCarte=$idCarte;
+      $posee->idJoueur=$idCarte;
+      $posee->save();
+      $possede = Possede::where('idCarte','=',$idCarte)
+                        ->where('idJoueur','=',$idJoueur)
+                        ->first();
+      if($possede->nbCartes==1){
+        $possede->forceDelete();
       }else{
-        $posee = Posee::where('idCarte','=',$idCarte)
-                      ->where('idJoueur','=',$idJoueur)
-                      ->first();
-        $posee->forceDelete();
+        $possede->nbCartes=$possede->nbCartes-1;
+        $possede->save();
       }
     }else{
       $manche =  Manche::where('idManche','=',$idManche)->first();
@@ -515,13 +472,7 @@ class ControlerPartie {
         $comporte->save();
       }
       $defausse = Defausse::where('idManche','=',$idManche)
-                          ->where('idJoueur','=',null)
                           ->first();
-      if($defausse==null){
-        $defausse->idManche=$idManche;
-        $defausse->idJoueur=null;
-        $defausse->save();
-      }
     }
     $estPlacee = EstPlacee::where('idDefausse','=',$defausse->idDefausse)
                           ->where('idCarte','=',$idCarte)
@@ -607,22 +558,16 @@ class ControlerPartie {
       }
       $possede->nbCartes++;
       $possede->save();
-      $possede2=Possede::where('idJoueur','=',$idJoueur)->first();
-      if(($possede2->idCarte==7 || $possede->idCarte==7) && ($possede2->idCarte==5 || $possede2->idCarte==6 || $possede->idCarte==5 || $possede->idCarte==6)){
+      $possede2=Possede::where('idJoueur','=',$idJoueur)
+                        ->where('idCarte','!=',$possede->idCarte)
+                        ->first();
+      if($possede2!=null && ($possede2->idCarte==7 || $possede->idCarte==7) && ($possede2->idCarte==5 || $possede2->idCarte==6 || $possede->idCarte==5 || $possede->idCarte==6)){
         $possede = Possede::where('idCarte','=',7)->first();
         $possede->forceDelete();
-        $defausse = Defausse::where('idJoueur','=',$_SESSION['idJoueur']);
-        if($defausse==null){
-          $defausse = new Defausse;
-          $defausse->idJoueur=$_SESSION['idJoueur'];
-          $defausse->idManche=$idManche;
-          $defausse->save();
-        }
-        $estPlacee = new EstPlacee;
-        $estPlacee->idCarte=7;
-        $estPlacee->idDefausse=$defausse->idDefausse;
-        $estPlacee->nbCartes=1;
-        $estPlacee->save();
+        $posee = new Posee;
+        $posee->idCarte=7;
+        $posee->idJoueur=$_SESSION['idJoueur'];
+        $posee->save();
       }
       $comporte=Comporte::where('idPioche',$idPioche)
                         ->where('idCarte', $randomId)
@@ -671,9 +616,13 @@ class ControlerPartie {
 
   public function eliminer($idJoueur){
     DBConnection::getInstance();
-    $joueur= Joueur::where('idJoueur','=',$idJoueur)->first();
+    $joueur = Joueur::where('idJoueur','=',$idJoueur)->first();
     $joueur->elimine=true;
     $joueur->save();
+    $possedes = Possede::where('idJoueur','=',$idJoueur)->get();
+    foreach($possedes as $possede){
+      $possede->forceDelete();
+    }
     $this->verifElimine();
   }
